+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 * Added new `AppMenuButton.renderWithUserProfile` prop as a built-in alternative to the default
   hamburger menu. Set to `true` to render the current user's initials instead or provide a function
   to render a custom element for the user.
+* Added `AggregationContext` as an additional argument to `CubeField.canAggregateFn`.
 
 ### ⚙️ Typescript API Adjustments